Bug fix: allow Mention autocomplete modal popup for every reply area
In older version, the auto complete modal pops up properly only with the top reply area. The reason was each reply area had its own modal and the modal in sub reply areas were too deep with the DOM tree that they cannot pop up as a top layer modal.

The bussiness logic is changed to using `Bus` as state & event manager to handle Mention autocomplete. And the modal tag is moved to `WfBody` (from `WfReplyArea`).