Added script for recording user data.
This script is used for getting websites that are currently using `wildfire`. The reason for collecting the data is that we are considering a major change to the project, and we hope we could be able to find all our users and inform them the change, to make sure of a smooth transition to the upcoming new version.